Ready for meeting later today
Got the various resource bars redisplaying, along with some pretty bad
collision detection, but enough to know that it works for objects in an
array. Still need to adjust health, energy, and shield and have effects
based on those...(this is something that needs to be connected to the
collision that Michael is working on) Also need to deal with random
events, but I think adding two of them shouldn't be too hard. Almost
there to a nice prototype guys!